src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["afonsolelis <afonso.brandao@prof.inteli.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.44"
numpy = "^2.1.2"
pandas = "^2.2.3"
jupyter = "^1.1.1"
pyarrow = "^17.0.0"
flask = "^3.0.3"
clickhouse-driver = "^0.2.9"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
flasgger = "^0.9.7.1"
pydantic-settings = "^2.6.1"
prefect = "^3.1.3"
streamlit = "^1.40.2"
plotly = "^5.24.1"
requests = "^2.32.3"
wordcloud = "^1.9.4"
matplotlib = "^3.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
